pyproject.toml
[tool.poetry]
name = "odyssey"
version = "0.1.0"
description = "Odyssey Dependencies"
authors = ["panhongqi <phq@prismslight.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.13.0"
aiomysql = "^0.0.20"
uvloop = "^0.14.0"
gunicorn = "^20.0"
fastapi = {version = "*", extras = ["all"]}
tortoise-orm = "*"
sentry-sdk = "^0.14.3"
ujson = "^2.0"
passlib = "^1.7"
PyJWT = "^1.7"
bcrypt = "^3.1"
click = "*"
raven = "*"
arq = "*"
redis = "*"
pycryptodomex = "*"
pyopenssl = "*"
httpx = "*"
BeautifulSoup4 = "^4.8"
lxml = "^4.4"
openpyxl = "*"
aerich = "*"
jinja2 = "*"

[tool.poetry.dev-dependencies]
taskipy = "*"
asynctest = "*"

[[tool.poetry.source]]
url = "https://mirrors.aliyun.com/pypi/simple/"
name = "aliyun"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
